requirements.txt
filelock==3.20.0
fsspec==2025.10.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
opencv-python==4.12.0.88
pillow==12.0.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
wheel==0.45.1
